requirements.txt
fastapi
uvicorn
streamlit
requests
pydantic
python-dotenv
groq
google-generativeai
openai
tenacity
httpx
aiohttp